requirements.txt
faker
groq
python-dotenv
tqdm
sentence-transformers
faiss-cpu
numpy
tiktoken